--- a/primefaces/src/main/resources/META-INF/resources/primefaces/core/core.utils.js
+++ b/primefaces/src/main/resources/META-INF/resources/primefaces/core/core.utils.js
@@ -1138,7 +1138,7 @@ if (!PrimeFaces.utils) {
                 }
             }
 
-            // Trigger onRemove events for widget.destroy and remove the element from the DOM and
+            // Trigger onRemove events for widget.destroy and remove the element from the DOM
             // IMPORTANT: This must occur before jq.off() to ensure the on("remove") events remain registered.
             if (removeElement) {
                 jq.triggerHandler("remove");