@cocorossello 2024-06-08
I can see that still with the patch the performance is related to options size. If I don't render the options is much faster (scripting time)

I think it is because the event bubbles up (maybe?). In any case, if you need the remove event it can be fired with:

```
    // Remove the element from the DOM and trigger onRemove events for widget.destroy.
    // IMPORTANT: This must occur before jq.off() to ensure the on("remove") events remain registered.
    if (removeElement) {
        jq.triggerHandler("remove");
        jq.get(0).remove()
    }
```

This way performance numbers are good:

```
AJAX updateElement : 100.050048828125 ms
common.js:41 Removed 22 DOM elements
common.js:40 AJAX updateElement : 1.617919921875 ms
common.js:41 Removed 1 DOM elements
common.js:40 AJAX updateElement : 44.89892578125 ms
common.js:41 Removed 2808 DOM elements
```
Can you test if this is enough for primefaces to cleanup?