@melloware 2024-06-09
@cocorossello yes let me test this in my real world scenarios