@melloware 2024-06-09
OK i did the same behavior in the app with current code and your patch.

```
BEFORE CHANGE:
VM43:121 AJAX updateElement: 75.097900390625 ms
VM43:122 Removed 138 DOM elements
VM43:121 AJAX updateElement: 211.8740234375 ms
VM43:122 Removed 199 DOM elements
VM43:121 AJAX updateElement: 302.350830078125 ms
VM43:122 Removed 11642 DOM elements
VM43:121 AJAX updateElement: 259.909912109375 ms
VM43:122 Removed 12050 DOM elements
VM43:553 TabView: 4.22998046875 ms
VM43:554 Removed 247 DOM elements
VM43:121 AJAX updateElement: 234.94091796875 ms
VM43:122 Removed 12020 DOM elements
VM43:553 TabView: 3.346923828125 ms
VM43:554 Removed 424 DOM elements
VM43:121 AJAX updateElement: 223.243896484375 ms
VM43:122 Removed 11596 DOM elements
VM43:553 TabView: 53.662109375 ms
VM43:554 Removed 11468 DOM elements
VM43:121 AJAX updateElement: 26.52294921875 ms
VM43:122 Removed 128 DOM elements


AFTER CHANGE:
VM43:122 AJAX updateElement: 73.114013671875 ms
VM43:123 Removed 138 DOM elements
VM43:122 AJAX updateElement: 212.7138671875 ms
VM43:123 Removed 199 DOM elements
VM43:122 AJAX updateElement: 281.281005859375 ms
VM43:123 Removed 11642 DOM elements
VM43:122 AJAX updateElement: 257.079833984375 ms
VM43:123 Removed 12050 DOM elements
VM43:554 TabView: 4.556884765625 ms
VM43:555 Removed 247 DOM elements
VM43:122 AJAX updateElement: 248.955078125 ms
VM43:123 Removed 12020 DOM elements
VM43:554 TabView: 6.14501953125 ms
VM43:555 Removed 424 DOM elements
VM43:122 AJAX updateElement: 211.81298828125 ms
VM43:123 Removed 11596 DOM elements
VM43:554 TabView: 54.3701171875 ms
VM43:555 Removed 11468 DOM elements
VM43:122 AJAX updateElement: 20.333984375 ms
VM43:123 Removed 128 DOM elements
```

in my testing it seems your change performs about the same but i can confirm all the memory is still being cleaned up.