@cocorossello 2024-06-09
Ok, I updated the PR with this findings. It still makes sense to skip the removal of select's children. (44ms skipping vs 130 without skipping)