@tandraschko 2024-06-10
+1 to merge when you added svg exclude